Updated main file, startup sequence, and commands.

+++ savedata.py
@@ -7,19 +7,41 @@ import json
 import errno
 import os
 
+# TODO: store data better
+version = 1.0
+authors = ["Jackson Elia", "Andrew Combs"]
+functionality = "Full Auto"
+
+visible = False
+timeout = 15
+
+username = ""
+password = ""
+
 
 class JSONManager(object):
     def __init__(self, fp: str=""):
+    
         if os.path.exists(fp): 
             self.fp = fp + "\save.json"
         else:
             raise FileNotFoundError(errno.ENOENT, os.strerror(errno.ENOENT), fp)
             return
         
-        self.buffer = {}
-        dump = json.dumps(self.buffer)
-        with open(self.fp, "w") as jsonfile:
-            jsonfile.write(dump)
+        # TODO: Have a better way of making sure that all options exist in the save file
+        if not os.path.exists(self.fp):    
+            self.buffer = {
+                "version": version,
+                "authors": authors,
+                "functionality": functionality,
+                "visible": visible,
+                "timeout": timeout,
+                "username": username,
+                "password": password,
+            }
+            dump = json.dumps(self.buffer)
+            with open(self.fp, "w") as jsonfile:
+                jsonfile.write(dump)
         
         return
